mqtt packid int16_t → uint16_t 全量同步（结构体/组包函数/绑定层/测试） path_trie 提取 _wildcard_must_own_seg / _path_scan_visit 消重；变量移至循环外 harbor 移除 timer_ctx wtimer 改用 nowsec()；补 wnow >= window_start 无符号下溢守卫 subcenter shared 投递 dsize/dbuf 移至循环外；行尾注释去列对齐 lcore _task_list_arg.n 与循环变量改 size_t 消 -Wsign-compare lcrypt digest __gc 补 secure_zero；hmac_final 输出缓冲补 secure_zero lmysql/lpgsql reader/stmt 释放后各方法补 NULL 守卫 ltask _ltask_register quecap 补负数校验 lprot custz_pack 失败多返 size=0；行尾注释压缩 lutils/lbson/lbytecache 变量移至循环外；注释压缩 Lua：srey.sleep 改 ms<=0；netread timeout 日志修正；websock 端口改按 url.scheme；redis nil 参数编为 $-1；pgsql.erro 补 or ""；http chunked 跳过空串

+++ lualib/lbind/lcore.c
@@ -8,7 +8,7 @@ typedef struct _task_entry {
     char name[64];
 }_task_entry;
 typedef struct _task_list_arg {
-    int32_t n;
+    size_t n;
     size_t cap;
     _task_entry *entries;
 }_task_list_arg;
@@ -614,7 +614,7 @@ static int32_t _lcore_task_list(lua_State *lua) {
     CALLOC(c.entries, c.cap, sizeof(_task_entry));
     loader_task_each(g_loader, _lcore_task_list_collect, &c);
     lua_newtable(lua);
-    int32_t i = 0;
+    size_t i = 0;
     for (; i < c.n; i++) {
         lua_newtable(lua);
         if ('\0' != c.entries[i].name[0]) {
